Fix 2x CodeQL cpp/integer-multiplication-cast-to-long warnings
- src/univalue/lib/univalue_write.cpp: cast prettyIndent to size_t
  before multiply (unsigned overflow before size_type conversion)
- src/crypto/muhash.cpp: cast d2 to double_limb_t before multiply
  (limb_t overflow before __int128 promotion, cf. lines 58/61/73)
CodeQL alerts #21 and #20.

+++ src/univalue/lib/univalue_write.cpp
@@ -63,7 +63,7 @@ std::string UniValue::write(unsigned int prettyIndent,
 
 static void indentStr(unsigned int prettyIndent, unsigned int indentLevel, std::string& s)
 {
-    s.append(prettyIndent * indentLevel, ' ');
+    s.append(static_cast<size_t>(prettyIndent) * indentLevel, ' ');
 }
 
 // NOLINTNEXTLINE(misc-no-recursion)
